ref: split up & simplified core actions
